requirements.txt


# Web App Framework (USED in app.py)
streamlit==1.30.0

# Machine Learning & Image Processing
numpy==1.26.2
pandas==2.1.4
scikit-learn==1.3.2
tensorflow==2.14.0
keras==2.14.0
opencv-python==4.8.1.78
pillow==10.1.0

# Firebase / Firestore
firebase-admin==6.3.0

# Language Translation & Detection
googletrans==4.0.0rc1
langdetect==1.0.9

# Voice Assistance (Text-to-Speech)
gTTS==2.5.1

# Data Visualization (used in dashboard)
matplotlib==3.8.2
seaborn==0.13.0

# API & Utilities
requests==2.31.0
python-dotenv==1.0.0
deep-translator

